requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-rest-framework==0.1.0
django-rest-knox==5.0.2
djangorestframework==3.16.1
sqlparse==0.5.3
